refactored comment tracking
