+++ app/helpers/application_helper.rb
@@ -46,26 +46,4 @@ module ApplicationHelper
   def auth_path
     '/auth/myusa'
   end
-
-  def markdown text
-    text = text ? text : ''
-    options = {
-      filter_html:     true,
-      hard_wrap:       true, 
-      link_attributes: { rel: 'nofollow', target: "_blank" },
-      space_after_headers: true, 
-      fenced_code_blocks: true
-    }
-
-    extensions = {
-      autolink:           true,
-      superscript:        true,
-      disable_indented_code_blocks: true
-    }
-
-    renderer = Redcarpet::Render::HTML.new(options)
-    markdown = Redcarpet::Markdown.new(renderer, extensions)
-
-    markdown.render(text).html_safe
-  end
 end
