+++ app/helpers/value_helper.rb
@@ -1,4 +1,5 @@
 module ValueHelper
+  include ActionView::Helpers::NumberHelper
   def date_with_tooltip(time)
     adjusted_time = time.in_time_zone("Eastern Time (US & Canada)").strftime("%b %-d, %Y at %l:%M%P")
     content_tag('span', adjusted_time, title: adjusted_time)
